Add files via upload
This version has a shortened name with new feedforward code (open loop) to proactively estimate what the valve balance should be using the hotBias parameter.  It also has NaN checking for all temperature channels and uses the last known good value for T0, T1, T2, T4, T5.   T3 is not used in this code as an active variable.